requirements.txt
discord.py>=2.6.0
googletrans==4.0.0rc1
python-dotenv
aiosqlite